moves document management to the cm file
